requirements.txt
# ===============================
# Core ML / LLM Stack
# ===============================
# transformers, accelerate, bitsandbytes, torch removed
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1

# ===============================
# LangChain Ecosystem (STABLE)
# ===============================
langchain==1.2.6
langchain-core==1.2.7
langchain-community==0.4.1
langchain-ollama==1.0.1
langgraph==1.0.7
langchain-openai>=0.2.0
openai>=1.12.0

# ===============================
# Vector Store
# ===============================
python-arango==8.1.3

# ===============================
# Privacy / HIPAA
# ===============================
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.7.4

# ===============================
# PDF / OCR / Multimodal
# ===============================
pymupdf4llm==0.2.9
pymupdf>=1.26.6
pytesseract==0.3.10
opencv-python==4.9.0.80
pillow==10.2.0

# ===============================
# API Layer
# ===============================
fastapi==0.128.0
uvicorn==0.27.1

# ===============================
# Utilities
# ===============================
tqdm==4.66.2
rich==13.7.0
python-dotenv==1.0.1
ipywidgets
pydantic>=2.9.2,<3.0.0
httpx==0.27.0
pydantic-settings>=2.0.0
structlog>=24.0.0
mcp>=1.0.0
orjson==3.10.1

# ===============================
# Database
# ===============================
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aioboto3>=12.3.0
python-multipart
